+++ packages/core/src/tracing/opentelemetry-instrumentations/index.js
@@ -7,5 +7,11 @@
 module.exports = {
   get otelInstrumentations() {
     return require('./wrap');
+  },
+  get preloadOtelInstrumentations() {
+    return require('./instrumentations').preloadOtelInstrumentations;
+  },
+  get getInstrumentationPackageNames() {
+    return require('./instrumentations').getInstrumentationPackageNames;
   }
 };
